+++ R/rbgm.R
@@ -13,8 +13,7 @@
 #' 
 #' @section I. Import: 
 #' \tabular{ll}{
-#'  \code{\link{bgm}} \tab read directly from a .bgm file  \cr
-#'  \code{\link{read.faces}} \tab read faces in with option to densify vertices \cr
+#'  \code{\link{bgmfile}} \tab read directly from a .bgm file  \cr
 #' }
 #' 
 #' @section II. Conversion: 
@@ -22,6 +21,8 @@
 #'  \code{\link{boxSpatial}} \tab convert boxes to a \code{\link{SpatialPolygonsDataFrame}}\cr
 #'  \code{\link{faceSpatial}} \tab convert faces to a \code{\link{SpatialLinesDataFrame}} \cr
 #'  \code{\link{boundarySpatial}} \tab convert boundary to a single-row \code{\link{SpatialPolygonsDataFrame}}\cr
+#'  \code{\link{nodeSpatial}} \tab obtain all vertices as points \cr
+#'  \code{\link{pointSpatial}} \tab obtain all instances of vertices as points \cr
 #' }
 #'
 #' @section III. Miscellaneous: